+++ core/src/elements/ons-list-item.js
@@ -15,7 +15,7 @@ limitations under the License.
 
 */
 
-import ons from '../ons';
+import onsElements from '../ons/elements';
 import util from '../ons/util';
 import styler from '../ons/styler';
 import autoStyle from '../ons/autostyle';
@@ -237,8 +237,8 @@ export default class ListItemElement extends BaseElement {
   }
 
   _onTouch(e) {
-    if (this.tapped || this !== e.target &&
-      (this._shouldIgnoreTap(e.target) || util.findParent(e.target, this._shouldIgnoreTap, p => p === this))
+    if (this.tapped ||
+      (this !== e.target && (this._shouldIgnoreTap(e.target) || util.findParent(e.target, this._shouldIgnoreTap, p => p === this)))
     ) {
       return; // Ignore tap
     }
@@ -265,5 +265,5 @@ export default class ListItemElement extends BaseElement {
   }
 }
 
-ons.elements.ListItem = ListItemElement;
+onsElements.ListItem = ListItemElement;
 customElements.define('ons-list-item', ListItemElement);