+++ core/src/elements/ons-progress-circular.js
@@ -15,7 +15,7 @@ limitations under the License.
 
 */
 
-import ons from '../ons';
+import onsElements from '../ons/elements';
 import util from '../ons/util';
 import autoStyle from '../ons/autostyle';
 import ModifierUtil from '../ons/internal/modifier-util';
@@ -236,5 +236,5 @@ export default class ProgressCircularElement extends BaseElement {
   }
 }
 
-ons.elements.ProgressCircular = ProgressCircularElement;
+onsElements.ProgressCircular = ProgressCircularElement;
 customElements.define('ons-progress-circular', ProgressCircularElement);