+++ core/src/elements/ons-segment.js
@@ -15,7 +15,7 @@ limitations under the License.
 
 */
 
-import ons from '../ons';
+import onsElements from '../ons/elements';
 import util from '../ons/util';
 import autoStyle from '../ons/autostyle';
 import ModifierUtil from '../ons/internal/modifier-util';
@@ -294,5 +294,5 @@ export default class SegmentElement extends BaseElement {
   }
 }
 
-ons.elements.Segment = SegmentElement;
+onsElements.Segment = SegmentElement;
 customElements.define('ons-segment', SegmentElement);