+++ core/src/elements/ons-switch.js
@@ -15,7 +15,7 @@ limitations under the License.
 
 */
 
-import ons from '../ons';
+import onsElements from '../ons/elements';
 import util from '../ons/util';
 import autoStyle from '../ons/autostyle';
 import ModifierUtil from '../ons/internal/modifier-util';
@@ -308,5 +308,5 @@ export default class SwitchElement extends BaseCheckboxElement {
 
 }
 
-ons.elements.Switch = SwitchElement;
+onsElements.Switch = SwitchElement;
 customElements.define('ons-switch', SwitchElement);